+++ cmd/livesim2/app/livesegment_test.go
@@ -88,22 +88,74 @@ func TestLiveSegment(t *testing.T) {
 			assert.Equal(t, tc.mediaTimescale, mediaTimescale)
 			media := tc.media
 			nr := 40
-			mediaTime := nr * 2 * mediaTimescale
+			mediaTime := nr * 2 * mediaTimescale // This is exact even for audio for nr == 40
 			switch mpdType {
 			case "Number", "TimelineNumber":
 				media = strings.Replace(media, "$NrOrTime$", fmt.Sprintf("%d", nr), -1)
 			default: // "TimelineTime":
 				media = strings.Replace(media, "$NrOrTime$", fmt.Sprintf("%d", mediaTime), -1)
 			}
-			segData, err := adjustLiveSegment(vodFS, asset, cfg, media, nowMS)
+			so, err := genLiveSegment(vodFS, asset, cfg, media, nowMS)
 			require.NoError(t, err)
-			require.Equal(t, tc.segmentMimeType, segData.contentType)
-			sr = bits.NewFixedSliceReader(segData.data)
-			mp4d, err = mp4.DecodeFileSR(sr)
-			require.NoError(t, err)
-			require.Equal(t, 1, len(mp4d.Segments), "nr segments should be 1")
-			bdt := mp4d.Segments[0].Fragments[0].Moof.Traf.Tfdt.BaseMediaDecodeTime()
+			require.Equal(t, tc.segmentMimeType, so.meta.rep.SegmentType())
+			seg := so.seg
+			bdt := seg.Fragments[0].Moof.Traf.Tfdt.BaseMediaDecodeTime()
 			require.Equal(t, mediaTime, int(bdt))
+		}
+	}
+}
+
+func TestCheckAudioSegmentTimeAddressing(t *testing.T) {
+	vodFS := os.DirFS("testdata/assets")
+	am := newAssetMgr(vodFS, "", false)
+	err := am.discoverAssets()
+	require.NoError(t, err)
+
+	cases := []struct {
+		asset        string
+		media        string
+		refSegDur    uint64
+		refTimescale uint64
+		nowMS        int
+		segNrStart   int
+		segNrEnd     int
+		nrSamplesMod []int
+	}{
+		{asset: "WAVE/vectors/cfhd_sets/14.985_29.97_59.94/t1/2022-10-17",
+			media: "A48/$NrOrTime$.m4s", refSegDur: 2002, refTimescale: 1000, nowMS: 50_000, segNrStart: 3, segNrEnd: 4,
+			nrSamplesMod: []int{94, 94, 94, 94, 94}},
+		{asset: "testpic_6s", media: "A48/$NrOrTime$.m4s", refSegDur: 6000, refTimescale: 1000, nowMS: 50_000, segNrStart: 3, segNrEnd: 7,
+			nrSamplesMod: []int{282, 281, 281, 281}},
+	}
+
+	for _, c := range cases {
+		for _, mpdType := range []string{"TimelineTime", "Number"} {
+			asset, ok := am.findAsset(c.asset)
+			require.True(t, ok)
+			cfg := NewResponseConfig()
+			switch mpdType {
+			case "Number":
+			case "TimelineTime":
+				cfg.SegTimelineFlag = true
+			case "TimelineNumber":
+				cfg.SegTimelineNrFlag = true
+			}
+			for nr := c.segNrStart; nr <= c.segNrEnd; nr++ {
+				mediaTime := calcAudioTimeFromRef(uint64(nr)*c.refSegDur, c.refTimescale, 1024, 48000)
+				var segMedia string
+				switch mpdType {
+				case "Number", "TimelineNumber":
+					segMedia = strings.Replace(c.media, "$NrOrTime$", fmt.Sprintf("%d", nr), -1)
+				default:
+					segMedia = strings.Replace(c.media, "$NrOrTime$", fmt.Sprintf("%d", mediaTime), -1)
+				}
+				so, err := genLiveSegment(vodFS, asset, cfg, segMedia, c.nowMS)
+				require.NoError(t, err)
+				trun := so.seg.Fragments[0].Moof.Traf.Trun
+				nrSamples := c.nrSamplesMod[nr%len(c.nrSamplesMod)]
+				require.Equal(t, nrSamples, len(trun.Samples))
+				fmt.Printf("nr %d segData: %s mpdType: %s mediaTime: %d\n", nr, so.meta.rep.SegmentType(), mpdType, mediaTime) // TODO. Remove
+			}
 		}
 	}
 }
@@ -137,7 +189,6 @@ func TestLiveThumbSegment(t *testing.T) {
 		for _, mpdType := range []string{"Number", "TimelineTime"} {
 			asset, ok := am.findAsset(tc.asset)
 			require.True(t, ok)
-			require.NoError(t, err)
 			cfg := NewResponseConfig()
 			switch mpdType {
 			case "Number":
@@ -150,13 +201,13 @@ func TestLiveThumbSegment(t *testing.T) {
 			media := tc.media
 			// Always number, even if MPD is timelinetime
 			media = strings.Replace(media, "$NrOrTime$", fmt.Sprintf("%d", tc.reqNr), -1)
-			segData, err := adjustLiveSegment(vodFS, asset, cfg, media, nowMS)
+			so, err := genLiveSegment(vodFS, asset, cfg, media, nowMS)
 			require.NoError(t, err)
 			origNr := tc.reqNr%tc.nrSegs + 1 // one-based
-			require.Equal(t, tc.segmentMimeType, segData.contentType)
+			require.Equal(t, tc.segmentMimeType, so.meta.rep.SegmentType())
 			origSeg, err := os.ReadFile(path.Join(tc.origPath, fmt.Sprintf("%d.jpg", origNr)))
 			require.NoError(t, err)
-			require.Equal(t, origSeg, segData.data)
+			require.Equal(t, origSeg, so.data)
 		}
 	}
 }
@@ -302,5 +353,71 @@ func TestTTMLTimeShifts(t *testing.T) {
 		gotTTML := string(gotTTMLBytes)
 		assert.Equal(t, c.wantedTTML, gotTTML)
 	}
+}
 
+func TestStartNumber(t *testing.T) {
+	vodFS := os.DirFS("testdata/assets")
+	am := newAssetMgr(vodFS, "", false)
+	err := am.discoverAssets()
+	require.NoError(t, err)
+	_, err = logging.InitZerolog("debug", "discard")
+	require.NoError(t, err)
+
+	cases := []struct {
+		asset              string
+		media              string
+		nowMS              int
+		startNr            int
+		requestNr          int
+		expectedDecodeTime int
+		expectedErr        string
+	}{
+		{
+			asset:              "testpic_2s",
+			media:              "V300/$NrOrTime$.m4s",
+			nowMS:              50_000,
+			startNr:            0,
+			requestNr:          0,
+			expectedDecodeTime: 0,
+			expectedErr:        "",
+		},
+		{
+			asset:       "testpic_2s",
+			media:       "V300/$NrOrTime$.m4s",
+			nowMS:       50_000,
+			startNr:     5,
+			requestNr:   0,
+			expectedErr: "createOutSeg: not found",
+		},
+		{
+			asset:              "testpic_2s",
+			media:              "A48/$NrOrTime$.m4s",
+			nowMS:              50_000,
+			startNr:            5,
+			requestNr:          5,
+			expectedDecodeTime: 0,
+			expectedErr:        "",
+		},
+	}
+	for _, tc := range cases {
+		asset, ok := am.findAsset(tc.asset)
+		require.True(t, ok)
+		require.NoError(t, err)
+		cfg := NewResponseConfig()
+		cfg.StartNr = Ptr(tc.startNr)
+		media := strings.Replace(tc.media, "$NrOrTime$", fmt.Sprintf("%d", (tc.requestNr)), 1)
+		so, err := genLiveSegment(vodFS, asset, cfg, media, tc.nowMS)
+		if tc.expectedErr != "" {
+			require.Error(t, err)
+			require.Contains(t, err.Error(), tc.expectedErr)
+			continue
+		}
+		require.NoError(t, err)
+		moof := so.seg.Fragments[0].Moof
+		seqNr := moof.Mfhd.SequenceNumber
+		require.Equal(t, tc.requestNr, int(seqNr), "response segment sequence number")
+		decodeTime := moof.Traf.Tfdt.BaseMediaDecodeTime()
+		require.Equal(t, tc.expectedDecodeTime, int(decodeTime), "response segment decode time")
+
+	}
 }