+++ CHANGELOG.md
@@ -15,13 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The online player is now proxied via /player when livesim2 runs with http
 - The online playURL should now including scheme
+- repdata (representation data on disk) format extended with commonSampleDuration
+- writerepdata option writes repdata even if existing
 
 ### Fixed
 
 - added muted=true to default playURL
 - HTTP 410 Gone response for segments before timeShiftBufferDepth
 - limited methods in OPTIONS response
-
 
 ## [0.8.0] - 2023-09-22